+++ source/data/iq_time_data.hh
@@ -24,11 +24,10 @@ namespace fast_daq
             virtual ~iq_time_data();
 
         public:
-            typedef double complex_t[2];
+            typedef float complex_t[2];
 
         // member varaible macros
         mv_accessible( unsigned, array_size );
-        //mv_accessible( std::vector< std::complex< double > >, volts_data );
         mv_accessible( complex_t*, data_array );
         mv_accessible( unsigned, chunk_counter );
