+++ source/data/real_time_data.cc
@@ -36,17 +36,16 @@ namespace fast_daq
         f_volts_data.resize( n_samples );
     }
 
-    std::vector<double> real_time_data::as_volts()
+    std::vector<float> real_time_data::as_volts()
     {
-         //std::vector<double> volts_data(f_array_size);
-         double units_factor = f_dynamic_range / 65536.;
-         double min_volts = f_dynamic_range / 2.0;
+         float units_factor = f_dynamic_range / 65536.;
+         float min_volts = f_dynamic_range / 2.0;
          for (unsigned i_bin=0; i_bin<f_array_size; ++i_bin)
          {
             //TODO validate this:
             // Note that volts = ((this_ADC_channel/number_of_channels) * dynamic_range) + min_voltage_in_range
             // ... where min_voltage_in_range == -(dynamic_range/2.)
-            f_volts_data[i_bin] = (static_cast<double>(f_time_series[i_bin]) * units_factor) - min_volts;
+            f_volts_data[i_bin] = (static_cast<float>(f_time_series[i_bin]) * units_factor) - min_volts;
          }
          return f_volts_data;
     }
